requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cloudinary==1.33.0
Django==4.2.2
django-ipware==5.0.0
gunicorn==20.1.0
mysqlclient==2.1.1
Pillow==9.5.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
